pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "keyspider"
version = "0.1.0"
description = "SSH key usage monitoring and tracking application"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"

dependencies = [
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "asyncssh>=2.14",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "celery[redis]>=5.3",
    "redis>=5.0",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "typer>=0.9",
    "rich>=13.0",
    "python-multipart>=0.0.6",
    "websockets>=12.0",
    "httpx>=0.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "httpx>=0.25",
    "aiosqlite>=0.19",
]

[project.scripts]
keyspider = "keyspider.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/keyspider"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
